Switch to using C++17
Summary
---

Partial backport from Core: https://github.com/bitcoin/bitcoin/pull/18591
And also ABC: https://reviews.bitcoinabc.org/D8493

Note that I also backported the `span.h` fixes from Pieter Wiulle in the
core#18591 backport, but I did not like way that class looked. It had
some unsafe behavior and was using `std::ptrdiff_t` instead of
`std::size_t`, among other problems.  I went ahead and spruced it up to
be more idiomatic and less insane.

    Commit message from Fabien:

    Debian buster ships with gcc 7/8 which is all green according to:
    https://gcc.gnu.org/projects/cxx-status.html

Test Plan
---

1. Build eveyrthing with cmake. Be sure to `rm -fr` your build directory
   and do a full build, run tests `ninja all check-all`
2. Build everything with the deprecated build system, again, using a
   fresh build dir: `./autogen.sh && ./configure --deprecated-build-system ...`
3. Run gitian build to make sure they are sane too.

+++ src/script/descriptor.cpp
@@ -522,7 +522,7 @@ enum class ParseScriptContext {
  * return true.
  */
 bool Const(const std::string &str, Span<const char> &sp) {
-    if ((size_t)sp.size() >= str.size() &&
+    if (sp.size() >= str.size() &&
         std::equal(str.begin(), str.end(), sp.begin())) {
         sp = sp.subspan(str.size());
         return true;
@@ -535,7 +535,7 @@ bool Const(const std::string &str, Span<const char> &sp) {
  * argument(s).
  */
 bool Func(const std::string &str, Span<const char> &sp) {
-    if ((size_t)sp.size() >= str.size() + 2 && sp[str.size()] == '(' &&
+    if (sp.size() >= str.size() + 2 && sp[str.size()] == '(' &&
         sp[sp.size() - 1] == ')' &&
         std::equal(str.begin(), str.end(), sp.begin())) {
         sp = sp.subspan(str.size() + 1, sp.size() - str.size() - 2);
